+++ assets/js/authentication/auth.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { withRouter, Redirect } from "react-router-dom";
-import { JSON_HEADERS } from "../constants/util_constants";
+import { JSON_HEADERS } from "../constants/requests";
 
 export const Authenticator = {
   isAuthenticated: !!localStorage.token,
