+++ assets/js/constants/requests.js
