+++ assets/js/productivity_log/add_productivity_event.js
@@ -1,6 +1,6 @@
 import Datetime from "react-datetime";
 import React, { Component, PropTypes } from "react";
-import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/util_constants";
+import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/requests";
 import moment from "moment";
 import { Link } from "react-router-dom";
 import { YEAR_MONTH_DAY_FORMAT } from "../constants/dates";
