+++ assets/js/resources_table/resource_table.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from "react";
-import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/util_constants";
+import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/requests";
 
 export class BaseEventLogTable extends Component {
   constructor() {
