+++ assets/js/sleep_log/add_sleep_event.js
@@ -2,7 +2,7 @@ import Datetime from "react-datetime";
 import React, { Component, PropTypes } from "react";
 import moment from "moment";
 import { Link } from "react-router-dom";
-import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/util_constants";
+import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/requests";
 
 export class AddSleepEvent extends Component {
   constructor(props) {
