+++ assets/js/user_activities_events_log/add_user_activities_event.js
@@ -3,7 +3,7 @@ import React, { Component, PropTypes } from "react";
 import {
   JSON_AUTHORIZATION_HEADERS,
   JSON_POST_AUTHORIZATION_HEADERS
-} from "../constants/util_constants";
+} from "../constants/requests";
 import moment from "moment";
 import { Link } from "react-router-dom";
 import { DASHBOARD_USER_ACTIVITIES_URL } from "../constants/urls";
