+++ assets/js/user_activities_log/add_user_activity.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from "react";
-import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/util_constants";
+import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/requests";
 import moment from "moment";
 import { Link } from "react-router-dom";
