+++ assets/js/user_settings/user_settings.js
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes } from "react";
-import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/util_constants";
+import { JSON_POST_AUTHORIZATION_HEADERS } from "../constants/requests";
 import { Authenticator } from "../authentication/auth";
 
 export class UserSettingsView extends Component {
